# Optimized memcpy implementation