**Implementation Detail:**
1. Converted src and destination pointer to word alligned
2. and then made word copy

**Performance Measurement:**
I used x86 64-bit ubuntu machine to compare the performance of new code with existing approach and gcc libc memcpy.
new code shows considerable improvement.
I also included inline x86 assembly approach (code is not included in review but its performace is even better than gcc libc memcpy).
Let me know if this can be used. I will include the code in review.

**Performance Analysis Report:**	 
Following input data used in evaluating performance with different approach:
Size selection: 64B, 128B, 256B, 512B, 1K, 2K ... 1M
Combination: alligned with 64 byte address, non-aligned destination,  non-aligned destination and src
Number of Loops: From 16777216 to 1024

Complete report is attached. 
[mempcy_performance_analysis.docx](https://github.com/nanovms/nanos/files/3370869/mempcy_performance_analysis.docx)

Following is one sample.
**memcpy 2048B -- 524288 loops**
  **aligned blocks with 64 multiple**
      nanos runtime_memcpy                               2.496000 s
      nanos runtime_memcpy_e                             0.640000 s
      libc memcpy                                        0.060000 s
      memcpy using x86 inline asssembly                  0.036000 s
  **unaligned blocks , src with +0 and dst with +4**
      nanos runtime_memcpy                               2.500000 s
      nanos runtime_memcpy_e                             0.648000 s
      libc memcpy                                        0.060000 s
      memcpy using x86 inline asssembly                  0.048000 s
  **unaligned blocks , src with +10 and dst with +3**
      nanos runtime_memcpy                               2.492000 s
      nanos runtime_memcpy_e                             0.764000 s
      libc memcpy                                        0.060000 s
      memcpy using x86 inline asssembly                  0.052000 s