@francescolavra 2019-07-17
I picked up this PR. I approved it because I had previously requested changes on it when it was being worked on by someone else, but this approval now doesn't count. So I'll wait for somebody to review it.