@francescolavra 2019-07-22
OK, I added to the unit test a memcpy test for all combinations of src_cnt and dest_cnt.